tdb --init tested

+++ tdb/manager.py
@@ -6,7 +6,8 @@ from string import Template
 __version__ = '0.0.1'
 
 
-code = Template('''from tdb.logger import GetLogger
+code = Template(\
+'''from tdb.logger import GetLogger
 from discord.ext.commands import Cog
 from discord.ext import commands
 from datetime import datetime
@@ -31,22 +32,51 @@ def setup(bot):
     bot.add_cog($Extension(bot))
 ''')
 
+code_main = \
+'''
+from tdb.bot import Main
+from dotenv import load_dotenv
+import os
+import tdb
+
+
+def get_config():
+    load_dotenv()
+    return dict(token=os.getenv('DEVELOP_TOKEN'))
+
+
+if __name__ == '__main__':
+    mybot = Main(get_config())
+    mybot.run()
+
+'''
+
+code_env = \
+'''
+DEVELOP_TOKEN=
+FINAL_TOKEN=
+'''
 
 def initworkspace():
     current_path = os.getcwd()
     extension_path = os.path.join(current_path, 'cogs')
     logs_path = os.path.join(current_path, 'logs')
     env_path = os.path.join(current_path, '.env')
+    main_path = os.path.join(current_path, 'main.py')
     if not os.path.exists(extension_path):
         print(f"Creating extension folder.")
         os.makedirs(extension_path)
     if not os.path.exists(logs_path):
         print(f"Creating logs folder.")
         os.makedirs(logs_path)
-    with open(env_path, 'w') as f:
-        print(f"Creating .env file.")
-        f.write('DEVELOP_TOKEN=')
-        f.write('FINAL_TOKEN=')
+    if not os.path.exists(env_path):
+        with open(env_path, 'w') as f:
+            print(f"Creating .env file.")
+            f.write(code_env)
+    if not os.path.exists(main_path):
+        with open(main_path, 'w') as f:
+            print(f"Creating main python file.")
+            f.write(code_main)
 
 
 
@@ -71,13 +101,13 @@ def remove_extension(Extension: str):
 
 def main():
         parser = argparse.ArgumentParser(description= f"Bot manager version {__version__}.")
-        parser.add_argument('--init',  nargs='?', default='None', type=str)
+        parser.add_argument('--init', action='store_true' , default=False)
         parser.add_argument('--new',  nargs='?', default='None', type=str)
         parser.add_argument('--delete',  nargs='?', default='None', type=str)
         config = parser.parse_args()
 
-        if config.init != 'None':
-            new_extension(config.new)
+        if config.init:
+            initworkspace()
 
         if config.new != 'None':
             new_extension(config.new)
